+++ src/hardrules/lm.py
@@ -261,13 +261,13 @@ def load_lm_filter(source_lang, target_lang, metadata_yaml, source_tokenizer_com
     lmFilter = DualLMFluencyFilter( LMType[metadata_yaml['lm_type']], source_lang, target_lang, source_tokenizer_command, target_tokenizer_command)
     stats=DualLMStats(metadata_yaml['clean_mean_perp'], metadata_yaml['clean_stddev_perp'], metadata_yaml['noisy_mean_perp'], metadata_yaml['noisy_stddev_perp'] )
 
-    fullpath_source_lm=os.path.join(metadata_yaml["yamlpath"], metadata_yaml['source_lm'])
+    fullpath_source_lm=os.path.join(metadata_yaml["yamldir"], metadata_yaml['source_lm'])
     if os.path.isfile(fullpath_source_lm):
         source_lm = fullpath_source_lm
     else:
         source_lm = metadata_yaml['source_lm']
 
-    fullpath_target_lm=os.path.join(metadata_yaml["yamlpath"], metadata_yaml['target_lm'])   
+    fullpath_target_lm=os.path.join(metadata_yaml["yamldir"], metadata_yaml['target_lm'])   
     if os.path.isfile(fullpath_target_lm):
         target_lm = fullpath_target_lm
     else: